Fix documentation, unify app-installation-token-related entities in a single file.
